+++ source/main/physics/RigSpawner.cpp
@@ -2664,7 +2664,7 @@ void ActorSpawner::ProcessRopable(RigDef::Ropable & def)
 {
     ropable_t ropable;
     ropable.node = GetNodePointerOrThrow(def.node);
-    ropable.pos = m_actor->ar_ropables.size();
+    ropable.pos = static_cast<int>(m_actor->ar_ropables.size());
     ropable.group = def.group;
     ropable.attached_ties = 0;
     ropable.attached_ropes = 0;
